+++ core-banking/src/main/java/com/example/corebanking/transfer/dto/TransferRequest.java
@@ -3,8 +3,10 @@ package com.example.corebanking.transfer.dto;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
+
 public record TransferRequest(
         @NotNull String toAccountNumber,
-        @Min(1) Long amount,
+        @Min(1) BigDecimal amount,
         @NotNull String transactionId
 ) {}
